Using classnames for conditional css
In this commit we created a function findUserLike
that receives all the likes from a post, filters
them comparing if the like at hand is from the
logged user. If so, the function returns true which
will be evaluated using classnames down below and
render or not the like button with a text-info
class applied to it.

Which makes the thumbs up button green or not.